feat: Prefer sections for maps when saving the initial ini file
BREAKING CHANGE: When the default ini file is saved, type-mapping and type-titles will be stord in sectoins with properties instead of with a single long property. The other way of representing a map - a string with colon and common delimiters - is still supported, it is just not the default one.