fix: Stop using underscores in configurable titles for notes like BREAKING_CHANGE
The ini parser was fixed to accept spaces in property names. There is no reason to replace " " with "_" any more. Original note names can (and have to) be used in the `type-titles` map to specify the titles of the chapters with these notes.

BREAKING CHANGE: If you configured the title of the breaking change notes using the key `BREAKING_CHANGE`, change it to `BREAKING CHANGE`. If you used other note sections with spaces in their names, do the same - use the same note name to configure the section title - with spaces.